A "small" rewrite of config subsystem.
And, yes, i still seriously disagree with your choice of config 
defaults, Bers.

Let it be recorded in commit history that i object against allowing 
people to be lazy. People should always know what they are dealing with, 
this is going to hit us hard one day.

When that day comes -- you will be the one rolling tech support.